+++ src/VisualStudio/Core/Def/TaskList/VisualStudioTaskListService.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Composition;
@@ -10,33 +11,45 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.Collections;
+using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.Editor.Shared.Utilities;
 using Microsoft.CodeAnalysis.Editor.TaskList;
 using Microsoft.CodeAnalysis.ErrorReporting;
 using Microsoft.CodeAnalysis.Host;
 using Microsoft.CodeAnalysis.Host.Mef;
 using Microsoft.CodeAnalysis.Options;
+using Microsoft.CodeAnalysis.PooledObjects;
 using Microsoft.CodeAnalysis.Shared.TestHooks;
+using Microsoft.CodeAnalysis.SolutionCrawler;
 using Microsoft.CodeAnalysis.TaskList;
 using Microsoft.VisualStudio.LanguageServices.Implementation.ProjectSystem;
 using Microsoft.VisualStudio.LanguageServices.Implementation.TableDataSource;
 using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Shell.Interop;
 using Microsoft.VisualStudio.Shell.TableManager;
+using Roslyn.Utilities;
 using TaskListItem = Microsoft.CodeAnalysis.TaskList.TaskListItem;
 
 namespace Microsoft.VisualStudio.LanguageServices.TaskList
 {
     [Export(typeof(VisualStudioTaskListService)), Shared]
-    internal class VisualStudioTaskListService : ITaskListProvider
+    internal class VisualStudioTaskListService : ITaskListProvider, ITaskListListener
     {
         private readonly IThreadingContext _threadingContext;
         private readonly VisualStudioWorkspaceImpl _workspace;
+        private readonly IGlobalOptionService _globalOptions;
         private readonly ITableManagerProvider _tableManagerProvider;
         private readonly IAsyncServiceProvider _asyncServiceProvider;
-        private readonly TaskListListener _listener;
 
         public event EventHandler<TaskListUpdatedArgs>? TaskListUpdated;
+
+        private readonly ConcurrentDictionary<DocumentId, ImmutableArray<TaskListItem>> _documentToTaskListItems = new();
+
+        /// <summary>
+        /// Queue where we enqueue the information we get from OOP to process in batch in the future.
+        /// </summary>
+        private readonly AsyncBatchingWorkQueue<(DocumentId documentId, ImmutableArray<TaskListItem> items)> _workQueue;
 
         [ImportingConstructor]
         [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
@@ -51,18 +64,14 @@ namespace Microsoft.VisualStudio.LanguageServices.TaskList
         {
             _threadingContext = threadingContext;
             _workspace = workspace;
+            _globalOptions = globalOptions;
             _tableManagerProvider = tableManagerProvider;
             _asyncServiceProvider = (IAsyncServiceProvider)asyncServiceProvider;
 
-            _listener = new TaskListListener(
-                globalOptions,
-                workspace.Services.SolutionServices,
-                asynchronousOperationListenerProvider,
-                onTaskListItemsUpdated: (documentId, oldComments, newComments) =>
-                {
-                    if (TaskListUpdated != null && !oldComments.SequenceEqual(newComments))
-                        TaskListUpdated?.Invoke(this, new TaskListUpdatedArgs(documentId, _workspace.CurrentSolution, documentId, newComments));
-                },
+            _workQueue = new AsyncBatchingWorkQueue<(DocumentId documentId, ImmutableArray<TaskListItem> items)>(
+                TimeSpan.FromSeconds(1),
+                ProcessTaskListItemsAsync,
+                asynchronousOperationListenerProvider.GetListener(FeatureAttribute.TaskList),
                 threadingContext.DisposalToken);
         }
 
@@ -92,7 +101,8 @@ namespace Microsoft.VisualStudio.LanguageServices.TaskList
                 // Now that we've started, create the actual VS todo list and have them hookup to us.
                 _ = new VisualStudioTaskListTable(workspace, _threadingContext, _tableManagerProvider, this);
 
-                _listener.Start();
+                // Now that we've hooked everything up, kick off the work to actually start computing and reporting items.
+                StartComputingTaskListItems();
             }
             catch (OperationCanceledException)
             {
@@ -136,6 +146,102 @@ namespace Microsoft.VisualStudio.LanguageServices.TaskList
         }
 
         public ImmutableArray<TaskListItem> GetTaskListItems(Workspace workspace, DocumentId documentId, CancellationToken cancellationToken)
-            => _listener.GetTaskListItems(documentId);
+            => _documentToTaskListItems.TryGetValue(documentId, out var values)
+                ? values
+                : ImmutableArray<TaskListItem>.Empty;
+
+        private void StartComputingTaskListItems()
+        {
+            // If we're in pull-diagnostics mode, then todo-comments will be handled by LSP.
+            var diagnosticMode = _globalOptions.GetDiagnosticMode();
+            if (diagnosticMode == DiagnosticMode.LspPull)
+                return;
+
+            var services = _workspace.Services.SolutionServices;
+            var registrationService = services.GetRequiredService<ISolutionCrawlerRegistrationService>();
+            var analyzerProvider = new TaskListIncrementalAnalyzerProvider(this);
+
+            registrationService.AddAnalyzerProvider(
+                analyzerProvider,
+                new IncrementalAnalyzerProviderMetadata(
+                    nameof(TaskListIncrementalAnalyzerProvider),
+                    highPriorityForActiveFile: false,
+                    workspaceKinds: WorkspaceKind.Host));
+        }
+
+        /// <summary>
+        /// Callback from the OOP service back into us.
+        /// </summary>
+        ValueTask ITaskListListener.ReportTaskListItemsAsync(DocumentId documentId, ImmutableArray<TaskListItem> items, CancellationToken cancellationToken)
+        {
+            try
+            {
+                _workQueue.AddWork((documentId, items));
+                return ValueTaskFactory.CompletedTask;
+            }
+            catch (Exception e) when (FatalError.ReportAndPropagateUnlessCanceled(e, cancellationToken))
+            {
+                // report NFW before returning back to the remote process
+                throw ExceptionUtilities.Unreachable();
+            }
+        }
+
+        /// <summary>
+        /// Callback from the OOP service back into us.
+        /// </summary>
+        ValueTask<TaskListOptions> ITaskListListener.GetOptionsAsync(CancellationToken cancellationToken)
+            => ValueTaskFactory.FromResult(_globalOptions.GetTaskListOptions());
+
+        private ValueTask ProcessTaskListItemsAsync(
+            ImmutableSegmentedList<(DocumentId documentId, ImmutableArray<TaskListItem> items)> docAndCommentsArray, CancellationToken cancellationToken)
+        {
+            cancellationToken.ThrowIfCancellationRequested();
+
+            using var _1 = ArrayBuilder<(DocumentId documentId, ImmutableArray<TaskListItem> items)>.GetInstance(out var filteredArray);
+            AddFilteredItems(docAndCommentsArray, filteredArray);
+
+            foreach (var (documentId, newItems) in filteredArray)
+            {
+                var oldComments = _documentToTaskListItems.TryGetValue(documentId, out var oldBoxedInfos)
+                    ? oldBoxedInfos
+                    : ImmutableArray<TaskListItem>.Empty;
+
+                // only one thread can be executing ProcessTodoCommentInfosAsync at a time,
+                // so it's safe to remove/add here.
+                if (newItems.IsEmpty)
+                {
+                    _documentToTaskListItems.TryRemove(documentId, out _);
+                }
+                else
+                {
+                    _documentToTaskListItems[documentId] = newItems;
+                }
+
+                // If we have someone listening for updates, and our new items are different from
+                // our old ones, then notify them of the change.
+                var taskListUpdated = this.TaskListUpdated;
+                if (TaskListUpdated != null && !oldComments.SequenceEqual(newItems))
+                    TaskListUpdated?.Invoke(this, new TaskListUpdatedArgs(documentId, _workspace.CurrentSolution, documentId, newItems));
+            }
+
+            return ValueTaskFactory.CompletedTask;
+        }
+
+        private static void AddFilteredItems(
+            ImmutableSegmentedList<(DocumentId documentId, ImmutableArray<TaskListItem> items)> array,
+            ArrayBuilder<(DocumentId documentId, ImmutableArray<TaskListItem> items)> filteredArray)
+        {
+            using var _ = PooledHashSet<DocumentId>.GetInstance(out var seenDocumentIds);
+
+            // Walk the list of todo comments in reverse, and skip any items for a document once
+            // we've already seen it once.  That way, we're only reporting the most up to date
+            // information for a document, and we're skipping the stale information.
+            for (var i = array.Count - 1; i >= 0; i--)
+            {
+                var info = array[i];
+                if (seenDocumentIds.Add(info.documentId))
+                    filteredArray.Add(info);
+            }
+        }
     }
 }
